Moved CommandWithSSH.php and class-terminus-command.php into php/Terminus/Commands, renamed the latter to TerminusCommand.php
