+++ php/Terminus/Commands/ArtCommand.php
@@ -1,6 +1,10 @@
 <?php
 
+namespace Terminus\Commands;
+
+use Terminus;
 use Terminus\Utils;
+use Terminus\Commands\TerminusCommand;
 use Terminus\Exceptions\TerminusException;
 
 /**
