+++ php/Terminus/Models/Environment.php
@@ -3,6 +3,7 @@
 namespace Terminus\Models;
 
 use Terminus\Request;
+use Terminus\Commands\TerminusCommand;
 use Terminus\Exceptions\TerminusException;
 use Terminus\Models\TerminusModel;
 use Terminus\Models\Collections\Bindings;
@@ -30,7 +31,7 @@ class Environment extends TerminusModel {
    * @return [array] $response['data']
    */
   public function addHostname($hostname) {
-    $response = \TerminusCommand::request(
+    $response = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       sprintf(
@@ -294,7 +295,7 @@ class Environment extends TerminusModel {
   public function create($env_name) {
     $path     = sprintf('environments/%s', $env_name);
     $params   = array();
-    $response = \TerminusCommand::request(
+    $response = TerminusCommand::request(
       'sites',
       $site_id,
       $path,
@@ -361,7 +362,7 @@ class Environment extends TerminusModel {
    * @return [array] $response['data']
    */
   public function deleteHostname($hostname) {
-    $response = \TerminusCommand::request(
+    $response = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       sprintf(
@@ -402,7 +403,7 @@ class Environment extends TerminusModel {
       'environments/%s/on-server-development/diffstat',
       $this->get('id')
     );
-    $data = \TerminusCommand::request(
+    $data = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
@@ -477,7 +478,7 @@ class Environment extends TerminusModel {
   public function getBackups($element = null) {
     if ($this->backups == null) {
       $path     = sprintf("environments/%s/backups/catalog", $this->get('id'));
-      $response = \TerminusCommand::request(
+      $response = TerminusCommand::request(
         'sites',
         $this->site->get('id'),
         $path,
@@ -521,7 +522,7 @@ class Environment extends TerminusModel {
       $element
     );
     $form_params = array('method' => 'GET');
-    $response    = \TerminusCommand::request(
+    $response    = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
@@ -541,7 +542,7 @@ class Environment extends TerminusModel {
       'environments/%s/on-server-development',
       $this->get('id')
     );
-    $result = \TerminusCommand::request(
+    $result = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
@@ -589,7 +590,7 @@ class Environment extends TerminusModel {
    * @return [array] $response['data']
    */
   public function getHostnames() {
-    $response = \TerminusCommand::request(
+    $response = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       'environments/' . $this->get('id') . '/hostnames',
@@ -616,7 +617,7 @@ class Environment extends TerminusModel {
    */
   public function info($key = null) {
     $path   = sprintf('environments/%s', $this->get('id'));
-    $result = \TerminusCommand::request(
+    $result = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
@@ -736,7 +737,7 @@ class Environment extends TerminusModel {
    */
   public function log() {
     $path     = sprintf('environments/%s/code-log', $this->get('id'));
-    $response = \TerminusCommand::request(
+    $response = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
@@ -858,7 +859,7 @@ class Environment extends TerminusModel {
       'type'        => $workflow,
       'environment' => $this->get('id'),
     );
-    $response    = \TerminusCommand::request(
+    $response    = TerminusCommand::request(
       'sites',
       $this->site->get('id'),
       $path,
