+++ php/class-terminus.php
@@ -29,6 +29,7 @@ class Terminus {
     $leaf_name = array_pop($path);
     $full_path = $path;
     $command   = self::getRootCommand();
+    $class     = "Terminus\\Commands\\$class";
 
     while (!empty($path)) {
       $subcommand_name = $path[0];
